Language, array, enum/flag, expression, primitive-spelling and API parity with dissect.cstruct (plan phases 0-5.1),
rebased onto 0.4.3. Details: CHANGELOG.md "Unreleased" and docs/guides/migrating-from-dissect.md.

+++ tests/CStructSharpTests/Enums.cs
@@ -21,7 +21,7 @@ public class Enums
 
         Assert.AreEqual("zing", enm.Name.Name);
         Assert.HasCount(3, enm.Values);
-        Assert.AreSame(Identifier.BYTE, enm.Type);
+        Assert.AreSame(Identifier.UINT32, enm.Type);
 
         Assert.AreEqual("Red", enm.Values[0].Name.Name);
         Assert.AreEqual(5, enm.Values[0].Value.Calc());
